Modified Reveal Average calculation. Average now given as a Fibonacci number

+++ backend/PlanningPoker.Api/DTOs/StatisticsDto.cs
@@ -2,7 +2,7 @@ namespace PlanningPoker.Api.DTOs;
 
 public class StatisticsDto
 {
-    public double Average { get; set; }
+    public int Average { get; set; }
     public int Min { get; set; }
     public int Max { get; set; }
 }
